QL: Remove the unnecessary `DateTimeFunction.dateTimeFormat()`
The `DateTimeFunction.dateTimeFormat()` method that was
called, but had no effect on the translated queries.

One of the examples, when this happens, is when one side of the
`BinaryComparison` is a function. The `BinaryComparison` at the
end turns into a `ScriptQuery`, but in the intermediate steps the
`ExpressionTranslators` try to translate it to a `RangeQuery`.

Follows #68783